requirements.txt
streamlit
python-docx
pandas
spacy
pymupdf
numpy
faiss-cpu
sentence-transformers
uuid
openai
python-dotenv
# tiktoken